+++ application/tests/models/Sites_Model_test.php
@@ -155,4 +155,16 @@ class Site_Model_test extends TestCase {
 		$this->assertRegExp('/^[0-9]+$/', $result['latest_chapter']);
 		$this->assertRegExp('/^[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+:[0-9]+$/', $result['last_updated']);
 	}
+	public function test_SeaOtterScans() {
+		$result = $this->Sites_Model->{'SeaOtterScans'}->getTitleData('marry_me');
+
+		$this->assertInternalType('array', $result);
+		$this->assertArrayHasKey('title', $result);
+		$this->assertArrayHasKey('latest_chapter', $result);
+		$this->assertArrayHasKey('last_updated', $result);
+
+		$this->assertEquals('Marry Me!', $result['title']);
+		$this->assertRegExp('/^[a-z]+\/[0-9]+\/[0-9]+(?:\/[0-9]+)?$/', $result['latest_chapter']);
+		$this->assertRegExp('/^[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+:[0-9]+$/', $result['last_updated']);
+	}
 }